requirements-retrieval.txt
datasets
pyarrow
jsonlines
tqdm
rank-bm25
faiss-cpu
sentence-transformers
numpy
scipy
scikit-learn